Revert "[Dakshta|Dhruvi] Resolve conflict"
This reverts commit e63af49bd1334617febdf3a4f23aaa72595b747a.
