+++ src/main/java/com/shopex/Message.java
@@ -1,0 +1,17 @@
+package com.shopex;
+
+public class Message {
+    private String message;
+
+    public Message(String message) {
+        this.message = message;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+    public void setMessage(String message) {
+        this.message = message;
+    }
+}
