+++ src/main/java/com/shopex/ShopexApplication.java
@@ -1,0 +1,13 @@
+package com.shopex;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ShopexApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ShopexApplication.class, args);
+	}
+
+}
