+++ src/main/java/com/shopex/exceptions/UserNameAlreadyExistsException.java
@@ -1,0 +1,7 @@
+package com.shopex.exceptions;
+
+public class UserNameAlreadyExistsException extends Exception {
+    public UserNameAlreadyExistsException() {
+        super("User name already exists");
+    }
+}
