+++ src/main/java/com/shopex/user/Role.java
@@ -1,0 +1,8 @@
+package com.shopex.user;
+
+public enum Role {
+    ADMIN,
+    USER;
+
+}
+
