+++ src/main/java/com/shopex/user/UserRepository.java
@@ -1,0 +1,11 @@
+package com.shopex.user;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    public Optional<User> findByUsername(String username);
+}
