--- src/main/java/com/tvastra/Message.java
@@ -1,17 +1,0 @@
-package com.tvastra;
-
-public class Message {
-    private String message;
-
-    public Message(String message) {
-        this.message = message;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-}
