--- src/main/java/com/tvastra/TvastraApplication.java
@@ -1,13 +1,0 @@
-package com.tvastra;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TvastraApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(TvastraApplication.class, args);
-	}
-
-}
