--- src/main/java/com/tvastra/config/AppConfig.java
@@ -1,22 +1,0 @@
-package com.tvastra.config;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@ConfigurationProperties(prefix = "application")
-public class AppConfig {
-
-    private String uiHost;
-
-    public String getUiHost() {
-        return uiHost;
-    }
-
-    public void setUiHost(String uiHost) {
-        this.uiHost = uiHost;
-    }
-
-}
-
-
