--- src/main/java/com/tvastra/config/MapperConfig.java
@@ -1,14 +1,0 @@
-package com.tvastra.config;
-
-import org.modelmapper.ModelMapper;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class MapperConfig {
-
-    @Bean
-    public ModelMapper modelMapper(){
-        return new ModelMapper();
-    }
-}
