--- src/main/java/com/tvastra/config/WebConfig.java
@@ -1,26 +1,0 @@
-package com.tvastra.config;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-@EnableWebMvc
-public class WebConfig implements WebMvcConfigurer {
-
-    private final AppConfig appConfig;
-
-    @Autowired
-    public WebConfig(AppConfig appConfig) {
-        this.appConfig = appConfig;
-    }
-
-    @Override
-    public void addCorsMappings(CorsRegistry registry) {
-        registry.addMapping("/**")
-                .allowedOrigins(appConfig.getUiHost())
-                .allowedMethods("*");
-    }
-}
