--- src/main/java/com/tvastra/exceptions/UserNameAlreadyExistsException.java
@@ -1,7 +1,0 @@
-package com.tvastra.exceptions;
-
-public class UserNameAlreadyExistsException extends Exception {
-    public UserNameAlreadyExistsException() {
-        super("User name already exists");
-    }
-}
