--- src/main/java/com/tvastra/gallery/Artist.java
@@ -1,65 +1,0 @@
-package com.tvastra.gallery;
-
-import com.tvastra.gallery.artwork.Artwork;
-
-import javax.persistence.*;
-import java.util.List;
-import java.util.Objects;
-
-@Entity
-public class Artist {
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    @Column(name = "id", nullable = false)
-    private Long id;
-    private String name;
-
-    @OneToMany(cascade = CascadeType.ALL, mappedBy = "artist")
-    private List<Artwork> artworks;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public List<Artwork> getArtworks() {
-        return artworks;
-    }
-
-    public void setArtworks(List<Artwork> artworks) {
-        this.artworks = artworks;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Artist artist = (Artist) o;
-        return Objects.equals(id, artist.id) && Objects.equals(name, artist.name) && Objects.equals(artworks, artist.artworks);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id, name, artworks);
-    }
-
-    @Override
-    public String toString() {
-        return "Artist{" +
-                "id=" + id +
-                ", name='" + name + '\'' +
-                ", artworks=" + artworks +
-                '}';
-    }
-}
