--- src/main/java/com/tvastra/gallery/category/CategoryRepository.java
@@ -1,8 +1,0 @@
-package com.tvastra.gallery.category;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CategoryRepository extends JpaRepository<Category, Long> {
-}
