--- src/main/java/com/tvastra/user/Role.java
@@ -1,8 +1,0 @@
-package com.tvastra.user;
-
-public enum Role {
-    ADMIN,
-    USER;
-
-}
-
