--- src/main/java/com/tvastra/user/UserRepository.java
@@ -1,10 +1,0 @@
-package com.tvastra.user;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    public Optional<User> findByUsername(String username);
-}
