+++ src/main/resources/application.properties
@@ -1,9 +1,4 @@
 spring.datasource.url=jdbc:postgresql://localhost:5432/tvastra
-<<<<<<< HEAD
 spring.datasource.username=${POSTGRES_USERNAME}
 spring.datasource.password=${POSTGRES_PASSWORD}
 application.uiHost=http://localhost:3000
-=======
-spring.datasource.username=${USERNAME}
-spring.datasource.password=${PASSWORD}
->>>>>>> 4c1b0b2 ([Dakshta|Dhruvi] Add. Artwork Database)
