--- src/main/resources/db/migration/V1.2__CreateCategoryTable.sql
@@ -1,5 +1,0 @@
-CREATE TABLE CATEGORY(
-    ID BIGINT PRIMARY KEY GENERATED BY DEFAULT AS IDENTITY,
-    NAME VARCHAR,
-    IMAGE_PATH VARCHAR
-);
