--- src/main/resources/db/migration/V1.3__CreateArtistTable.sql
@@ -1,4 +1,0 @@
-CREATE TABLE ARTIST (
-    ID BIGINT PRIMARY KEY GENERATED BY DEFAULT AS IDENTITY,
-    NAME VARCHAR
-);
