--- src/main/resources/db/migration/V1.4__CreateGalleryTable.sql
@@ -1,6 +1,0 @@
-CREATE TABLE GALLERY (
-    ID BIGINT PRIMARY KEY GENERATED BY DEFAULT AS IDENTITY,
-    NAME VARCHAR,
-    OWNER_ID BIGINT,
-    CONSTRAINT fk_gallery_user FOREIGN KEY (OWNER_ID) REFERENCES USERS (ID)
-);
