--- src/main/resources/db/migration/V1.5__CreateArtworkTable.sql
@@ -1,10 +1,0 @@
-CREATE TABLE ARTWORK(
-    ID BIGINT PRIMARY KEY GENERATED BY DEFAULT AS IDENTITY,
-    TITLE VARCHAR,
-    DESCRIPTION VARCHAR,
-    ARTIST_ID BIGINT,
-    GALLERY_ID BIGINT,
-    PRICE DECIMAL,
-    CONSTRAINT fk_artwork_artist FOREIGN KEY (ARTIST_ID) REFERENCES ARTIST (ID),
-    CONSTRAINT fk_artwork_gallery FOREIGN KEY (GALLERY_ID) REFERENCES GALLERY (ID)
-);
