+++ src/test/java/com/shopex/ShopexApplicationTests.java
@@ -1,0 +1,13 @@
+//package com.shopex;
+//
+//import org.junit.jupiter.api.Test;
+//import org.springframework.boot.test.context.SpringBootTest;
+//
+//@SpringBootTest
+//class ShopexApplicationTests {
+//
+//	@Test
+//	void contextLoads() {
+//	}
+//
+//}
