--- src/test/java/com/tvastra/TvastraApplicationTests.java
@@ -1,13 +1,0 @@
-package com.tvastra;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TvastraApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
